+++ src/components/orchard-details/field-info-section.tsx
@@ -4,7 +4,7 @@ import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
 import { Separator } from "@/components/ui/separator";
 import { Combobox } from "../combobox";
-import { comboBoxOptions } from "@/lib/utils";
+import { comboBoxOptions } from "@/lib/utils/comboBoxOptions";
 
 interface FieldInfoSectionProps {
   formData: {
