+++ src/components/orchard-details/team-info-section.tsx
@@ -3,7 +3,7 @@
 import { Label } from "@/components/ui/label";
 import { Separator } from "@/components/ui/separator";
 import { Combobox } from "../combobox";
-import { comboBoxOptions } from "@/lib/utils";
+import { comboBoxOptions } from "@/lib/utils/comboBoxOptions";
 
 interface TeamInfoSectionProps {
   formData: {
