+++ src/components/ui/accordion.tsx
@@ -3,7 +3,7 @@ import * as AccordionPrimitive from "@radix-ui/react-accordion"
 import { ChevronDownIcon } from "lucide-react"
 import { CORNERS } from "@/constants"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/cn"
 
 function Accordion({
   ...props
