+++ src/lib/utils/env-validation.ts
